Reorganized project structure and adjusted names (we're using entities not some domain model classes so our naming convention should reflect that)
