+++ mvc/Configurations/Entities/QuoteConfiguration.cs
@@ -1,35 +1,35 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using mvc.Models;
+using mvc.DataAccess.Entities;
 
 namespace mvc.Configurations.Entities;
 
-public class QuoteConfiguration : IEntityTypeConfiguration<Quote>
+public class QuoteConfiguration : IEntityTypeConfiguration<QuoteEntity>
 {
-    public void Configure(EntityTypeBuilder<Quote> builder)
+    public void Configure(EntityTypeBuilder<QuoteEntity> builder)
     {
         builder.HasData(
-            new Quote
+            new QuoteEntity
             {
                 Id = 1,
                 Text = "Per aspera ad astra.",
                 Author = "Unknown",
                 Language = "Latin"
             },
-            new Quote
+            new QuoteEntity
             {
                 Id = 2,
                 Text = "Jestem życiem, które pragnie żyć, pośród życia, które pragnie żyć.",
                 Author = "Albert Shweitzer",
                 Language = "Polish"
-            }, new Quote
+            }, new QuoteEntity
             {
                 Id = 3,
                 Text = "Never be the first, never be the last, and never ever volunteer to do anything.",
                 Author = "Military Wisdom",
                 Language = "English"
             },
-            new Quote
+            new QuoteEntity
             {
                 Id = 4,
                 Text = "Niczego się nie wyrzekać, do niczego się nie przywiązywać.",
