+++ mvc/Configurations/Mapper/QuoteProfile.cs
@@ -1,6 +1,6 @@
 ﻿using AutoMapper;
+using mvc.DataAccess.Entities;
 using mvc.DTOs;
-using mvc.Models;
 
 namespace mvc.Configurations.Mapper;
 
@@ -8,8 +8,8 @@ public class QuoteProfile : Profile
 {
     public QuoteProfile()
     {
-        CreateMap<Quote, QuoteDto>().ReverseMap();
-        CreateMap<QuoteForCreationDto, Quote>().ReverseMap();
-        CreateMap<QuoteForUpdateDto, Quote>().ReverseMap();
+        CreateMap<QuoteEntity, QuoteDto>().ReverseMap();
+        CreateMap<QuoteForCreationDto, QuoteEntity>().ReverseMap();
+        CreateMap<QuoteForUpdateDto, QuoteEntity>().ReverseMap();
     }
 }
