+++ mvc/DataAccess/Repository/Interfaces/IQuoteRepository.cs
@@ -1,10 +1,10 @@
-﻿using mvc.Models;
+﻿using mvc.DataAccess.Entities;
 
 namespace mvc.DataAccess.Repository.Interfaces;
 
-public interface IQuoteRepository : IRepository<Quote>
+public interface IQuoteRepository : IRepository<QuoteEntity>
 {
-    Task<(IReadOnlyList<Quote>, PaginationMetadata)> GetQuotesAsync(
+    Task<(IReadOnlyList<QuoteEntity>, PaginationMetadata)> GetQuotesAsync(
         string? author,
         string? language,
         string? text,
